requirements.txt
fastapi==0.115.0
uvicorn==0.30.0
chromadb==0.5.0
sentence-transformers==3.0.0
pypdf==4.0.0
python-docx==1.1.0
aiofiles==24.1.0
python-multipart==0.0.9
httpx==0.27.0
